requirements.txt
appdirs==1.4.4
attrs==24.2.0
blinker==1.9.0
click==8.1.7
customtkinter==5.2.2
darkdetect==0.8.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
jawa==2.2.0
Jinja2==3.1.4
jsonlines==4.0.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycryptodome==3.21.0
Pygments==2.18.0
pyshark==0.6
python-dateutil==2.9.0.post0
pytz==2024.2
RestrictedPython==7.4
rich==13.9.4
six==1.17.0
termcolor==2.5.0
tkinterhtml==0.7
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
zipp==3.21.0
